requirements.txt
Django==1.10.3
Pillow>=2.0.0
Django-ckeditor==5.1.1
djangorestframework==3.5.3
djangorestframework-camel-case==0.2.0
django_cors_headers==1.3.1